refactor: Change stock fields from decimal to integer in StockProduct model and migration. This update modifies the data type of min_stock and max_stock fields to integer for better alignment with stock management requirements, ensuring consistency across the model and database migrations.

+++ database/migrations/2026_01_20_111221_add_image_and_stock_fields_to_stock_products_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
     {
         Schema::table('stock_products', function (Blueprint $table) {
             $table->string('image_path', 500)->nullable()->after('active');
-            $table->decimal('min_stock', 10, 4)->nullable()->after('image_path');
-            $table->decimal('max_stock', 10, 4)->nullable()->after('min_stock');
+            $table->integer('min_stock')->nullable()->after('image_path');
+            $table->integer('max_stock')->nullable()->after('min_stock');
         });
     }
